Add tests to support -S/--gpg-sign/--no-gpg-sign command line options and commit.gpgsign configuration.
These are currently implemented as a copy of t7900-subtree.sh,
and do not pass completely (4 of 58 fail).
Followup commits will fix these last few tests and combine
with the existing subtree tests.

t/test-lib-functions.sh test_commit: Add support for passing --gpg-sign to test_commit.

This change adds support for passing
--gpg-sign, --no-gpg-sign, and -S to the test_commit function

When gpg signing support is added to contrib/subtree, this will be used
to sign the test commits.

Signed-off-by: Allen Reese <allen_reese@apple.com>